requirements.txt
lxml
PyExecJS
pytesseract
requests
Pillow
flask
